@@ -4,15 +4,16 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_screenutil: ^0.5.2
-  auto_size_text: ^2.1.0
-  flutter_masked_text: ^0.8.0
-  cupertino_icons: ^0.1.2
+  flutter_screenutil: ^5.8.4
+  auto_size_text: ^3.0.0
+  flutter_masked_text2: ^0.9.1
+  cupertino_icons: ^1.0.5
+
 
 dev_dependencies:
   flutter_test:
