@@ -17,19 +17,20 @@ class PageCodigoDeBarraV1 extends StatelessWidget {
 }
 
 class CodigoDeBarrav1 extends StatefulWidget {
-  CodigoDeBarrav1({Key key, this.title}) : super(key: key);
-  final String title;
+  CodigoDeBarrav1({Key? key, this.title}) : super(key: key);
+  final String? title;
   _CodigoDeBarrav1 createState() => _CodigoDeBarrav1();
 }
 
 class _CodigoDeBarrav1 extends State<CodigoDeBarrav1> {
   static const platform = const MethodChannel('samples.flutter.dev/gedi');
   final myController = TextEditingController();
-  List<String> resultadosBardCod = new List<String>();
+  List<String> resultadosBardCod = [];
 
   Future<void> _leitorCodigoDeBarra(String tipoLeitura) async {
     try {
-      String teste = await platform.invokeMethod('leitorCodigov1', <String, dynamic>{"tipoLeitura": tipoLeitura});
+      String teste = await platform.invokeMethod(
+          'leitorCodigov1', <String, dynamic>{"tipoLeitura": tipoLeitura});
       setState(() {
         resultadosBardCod.add(teste);
       });
@@ -41,32 +42,35 @@ class _CodigoDeBarrav1 extends State<CodigoDeBarrav1> {
   Widget build(BuildContext context) {
     double defaultScreenWidth = 400.0;
     double defaultScreenHeight = 810.0;
-    ScreenUtil.instance = ScreenUtil(
-      width: defaultScreenWidth,
-      height: defaultScreenHeight,
-      allowFontScaling: true,
-    )..init(context);
+    ScreenUtil.init(
+      context,
+      designSize: Size(defaultScreenWidth, defaultScreenHeight),
+    );
 
     return new Scaffold(
       body: Container(
         padding: EdgeInsets.only(
-          top: ScreenUtil.instance.setWidth(40.0),
+          top: ScreenUtil().setWidth(40.0),
         ),
         child: Column(
           children: <Widget>[
             Text(
               "Ler Código de Barras",
-              style: TextStyle(color: Colors.black, fontWeight: FontWeight.bold, fontSize: 30),
+              style: TextStyle(
+                  color: Colors.black,
+                  fontWeight: FontWeight.bold,
+                  fontSize: 30),
             ),
             Row(
               mainAxisAlignment: MainAxisAlignment.spaceEvenly,
               children: [
                 SizedBox(
-                  width: ScreenUtil.instance.setWidth(100),
-                  child: RaisedButton(
+                  width: ScreenUtil().setWidth(100),
+                  child: ElevatedButton(
                     child: Text(
                       'EAN 8',
-                      style: TextStyle(fontSize: ScreenUtil.instance.setSp(15), color: Colors.red),
+                      style: TextStyle(
+                          fontSize: ScreenUtil().setSp(15), color: Colors.red),
                     ),
                     onPressed: () {
                       _leitorCodigoDeBarra("EAN_8");
@@ -74,11 +78,12 @@ class _CodigoDeBarrav1 extends State<CodigoDeBarrav1> {
                   ),
                 ),
                 SizedBox(
-                  width: ScreenUtil.instance.setWidth(100),
-                  child: RaisedButton(
+                  width: ScreenUtil().setWidth(100),
+                  child: ElevatedButton(
                     child: Text(
                       'EAN 13',
-                      style: TextStyle(fontSize: ScreenUtil.instance.setSp(15), color: Colors.red),
+                      style: TextStyle(
+                          fontSize: ScreenUtil().setSp(15), color: Colors.red),
                     ),
                     onPressed: () {
                       _leitorCodigoDeBarra("EAN_13");
@@ -86,11 +91,12 @@ class _CodigoDeBarrav1 extends State<CodigoDeBarrav1> {
                   ),
                 ),
                 SizedBox(
-                  width: ScreenUtil.instance.setWidth(100), // specific value
-                  child: RaisedButton(
+                  width: ScreenUtil().setWidth(100), // specific value
+                  child: ElevatedButton(
                     child: Text(
                       'EAN 14',
-                      style: TextStyle(fontSize: ScreenUtil.instance.setSp(15), color: Colors.red),
+                      style: TextStyle(
+                          fontSize: ScreenUtil().setSp(15), color: Colors.red),
                     ),
                     onPressed: () {
                       _leitorCodigoDeBarra("EAN_14");
@@ -98,11 +104,12 @@ class _CodigoDeBarrav1 extends State<CodigoDeBarrav1> {
                   ),
                 ),
                 SizedBox(
-                  width: ScreenUtil.instance.setWidth(100), // specific value
-                  child: RaisedButton(
+                  width: ScreenUtil().setWidth(100), // specific value
+                  child: ElevatedButton(
                     child: Text(
                       'QR CODE',
-                      style: TextStyle(fontSize: ScreenUtil.instance.setSp(15), color: Colors.red),
+                      style: TextStyle(
+                          fontSize: ScreenUtil().setSp(15), color: Colors.red),
                     ),
                     onPressed: () {
                       _leitorCodigoDeBarra("QR_CODE");
@@ -123,7 +130,10 @@ class _CodigoDeBarrav1 extends State<CodigoDeBarrav1> {
                       dense: true,
                       title: Text(
                         resultadosBardCod[index],
-                        style: TextStyle(fontWeight: FontWeight.bold, fontSize: ScreenUtil.instance.setSp(12.0), color: Colors.black54),
+                        style: TextStyle(
+                            fontWeight: FontWeight.bold,
+                            fontSize: ScreenUtil().setSp(12.0),
+                            color: Colors.black54),
                       ),
                     ),
                   );
