@@ -18,8 +18,8 @@ class PageImprimir extends StatelessWidget {
 }
 
 class Imprimir extends StatefulWidget {
-  Imprimir({Key key, this.title}) : super(key: key);
-  final String title;
+  Imprimir({Key? key, this.title}) : super(key: key);
+  final String? title;
   _Imprimir createState() => _Imprimir();
 }
 
@@ -29,12 +29,12 @@ class _Imprimir extends State<Imprimir> {
   int dropdownValueSize = 20;
   int dropdownBarHeight = 280;
   int dropdownBarWidth = 280;
-  String dropdownFont = "DEFAULT";
-  String dropdownBarType = "QR_CODE";
-  String alinhar = "CENTER";
-  String valorSelecionado = "CENTER";
+  String? dropdownFont = "DEFAULT";
+  String? dropdownBarType = "QR_CODE";
+  String? alinhar = "CENTER";
+  String? valorSelecionado = "CENTER";
   List<bool> listSelecionado = [false, false, false];
-  void radioButtonChanges(String value) {
+  void radioButtonChanges(String? value) {
     setState(
       () {
         valorSelecionado = value;
@@ -72,7 +72,13 @@ class _Imprimir extends State<Imprimir> {
     await platform.invokeMethod('fimimpressao');
   }
 
-  void impressaoDeTexto({String texto, int fontSize, String alinhar, String fontFamily, List<bool> selectedOptions}) async {
+  void impressaoDeTexto({
+    String? texto,
+    int? fontSize,
+    String? alinhar,
+    String? fontFamily,
+    List<bool>? selectedOptions,
+  }) async {
     texto = texto ?? ""; // Caso seja null
     if (texto.isEmpty) {
       erroImpresao();
@@ -118,7 +124,8 @@ class _Imprimir extends State<Imprimir> {
     }
   }
 
-  Future<void> _impressaoDeCodigoDeBarra({String texto, int height, int width, String barCode}) async {
+  Future<void> _impressaoDeCodigoDeBarra(
+      {String? texto, int? height, int? width, String? barCode}) async {
     texto = texto ?? ""; // Caso seja null
     if (texto.isNotEmpty) {
       try {
@@ -153,11 +160,10 @@ class _Imprimir extends State<Imprimir> {
   Widget build(BuildContext context) {
     double defaultScreenWidth = 400.0;
     double defaultScreenHeight = 810.0;
-    ScreenUtil.instance = ScreenUtil(
-      width: defaultScreenWidth,
-      height: defaultScreenHeight,
-      allowFontScaling: true,
-    )..init(context);
+    ScreenUtil.init(
+      context,
+      designSize: Size(defaultScreenWidth, defaultScreenHeight),
+    );
 
     return new Scaffold(
       body: Center(
@@ -174,13 +180,15 @@ class _Imprimir extends State<Imprimir> {
                   fit: BoxFit.fitWidth,
                   child: Text(
                     "Funções Impressão G700/G800",
-                    style: TextStyle(fontSize: ScreenUtil.instance.setSp(25), fontWeight: FontWeight.bold),
+                    style: TextStyle(
+                        fontSize: ScreenUtil().setSp(25),
+                        fontWeight: FontWeight.bold),
                   ),
                 ),
               ),
               SizedBox(
                 width: 300,
-                child: RaisedButton(
+                child: ElevatedButton(
                   child: Text("STATUS IMPRESSORA"),
                   onPressed: () {
                     checarImpressora().then((value) {
@@ -211,24 +219,35 @@ class _Imprimir extends State<Imprimir> {
                   fit: BoxFit.fitWidth,
                   child: Text(
                     "Configuração de Impressão",
-                    style: TextStyle(fontSize: ScreenUtil.instance.setSp(25), fontWeight: FontWeight.bold),
+                    style: TextStyle(
+                        fontSize: ScreenUtil().setSp(25),
+                        fontWeight: FontWeight.bold),
                   ),
                 ),
               ),
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: <Widget>[
-                  Radio(value: 'LEFT', groupValue: valorSelecionado, onChanged: radioButtonChanges),
+                  Radio(
+                      value: 'LEFT',
+                      groupValue: valorSelecionado,
+                      onChanged: radioButtonChanges),
                   AutoSizeText(
                     'Esquerda',
                     style: TextStyle(fontSize: 15),
                   ),
-                  Radio(value: 'CENTER', groupValue: valorSelecionado, onChanged: radioButtonChanges),
+                  Radio(
+                      value: 'CENTER',
+                      groupValue: valorSelecionado,
+                      onChanged: radioButtonChanges),
                   AutoSizeText(
                     'Centralizado',
                     style: TextStyle(fontSize: 15),
                   ),
-                  Radio(value: 'RIGHT', groupValue: valorSelecionado, onChanged: radioButtonChanges),
+                  Radio(
+                      value: 'RIGHT',
+                      groupValue: valorSelecionado,
+                      onChanged: radioButtonChanges),
                   AutoSizeText(
                     'Direita',
                     style: TextStyle(fontSize: 15),
@@ -239,9 +258,12 @@ class _Imprimir extends State<Imprimir> {
                 mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                 children: <Widget>[
                   SizedBox(
-                    width: ScreenUtil.instance.setWidth(110),
-                    child: RaisedButton(
-                      color: listSelecionado[0] ? Colors.blue : Colors.grey,
+                    width: ScreenUtil().setWidth(110),
+                    child: ElevatedButton(
+                      style: ButtonStyle(
+                        backgroundColor: MaterialStateProperty.all<Color>(
+                            listSelecionado[0] ? Colors.blue : Colors.grey),
+                      ),
                       onPressed: () {
                         setState(() {
                           listSelecionado[0] = !listSelecionado[0];
@@ -249,39 +271,43 @@ class _Imprimir extends State<Imprimir> {
                       },
                       child: Text(
                         "NEGRITO",
-                        style: TextStyle(fontSize: ScreenUtil.instance.setSp(12)),
+                        style: TextStyle(fontSize: ScreenUtil().setSp(12)),
                       ),
                     ),
                   ),
                   SizedBox(
-                    width: ScreenUtil.instance.setWidth(110),
-                    child: RaisedButton(
-                      color: listSelecionado[1] ? Colors.blue : Colors.grey,
+                    width: ScreenUtil().setWidth(110),
+                    child: ElevatedButton(
+                      style: ButtonStyle(
+                        backgroundColor: MaterialStateProperty.all<Color>(
+                            listSelecionado[0] ? Colors.blue : Colors.grey),
+                      ),
                       onPressed: () {
                         setState(() {
                           listSelecionado[1] = !listSelecionado[1];
                         });
                       },
-                      disabledTextColor: Colors.black87,
                       child: Text(
                         "ITÁLICO",
-                        style: TextStyle(fontSize: ScreenUtil.instance.setSp(12)),
+                        style: TextStyle(fontSize: ScreenUtil().setSp(12)),
                       ),
                     ),
                   ),
                   SizedBox(
-                    width: ScreenUtil.instance.setWidth(110),
-                    child: RaisedButton(
-                      color: listSelecionado[2] ? Colors.blue : Colors.grey,
+                    width: ScreenUtil().setWidth(110),
+                    child: ElevatedButton(
+                      style: ButtonStyle(
+                        backgroundColor: MaterialStateProperty.all<Color>(
+                            listSelecionado[0] ? Colors.blue : Colors.grey),
+                      ),
                       onPressed: () {
                         setState(() {
                           listSelecionado[2] = !listSelecionado[2];
                         });
                       },
-                      disabledTextColor: Colors.black87,
                       child: Text(
                         "SUBLINHADO",
-                        style: TextStyle(fontSize: ScreenUtil.instance.setSp(12)),
+                        style: TextStyle(fontSize: ScreenUtil().setSp(12)),
                       ),
                     ),
                   )
@@ -299,10 +325,13 @@ class _Imprimir extends State<Imprimir> {
                     icon: Icon(Icons.arrow_downward),
                     iconSize: 24,
                     elevation: 16,
-                    style: TextStyle(color: Colors.black, fontWeight: FontWeight.bold, fontSize: 20),
-                    onChanged: (String newValue) {
+                    style: TextStyle(
+                        color: Colors.black,
+                        fontWeight: FontWeight.bold,
+                        fontSize: 20),
+                    onChanged: (String? newValue) {
                       setState(() {
-                        dropdownFont = newValue;
+                        dropdownFont = newValue!;
                       });
                     },
                     items: <String>[
@@ -327,13 +356,17 @@ class _Imprimir extends State<Imprimir> {
                     icon: Icon(Icons.arrow_downward),
                     iconSize: 24,
                     elevation: 16,
-                    style: TextStyle(color: Colors.black, fontWeight: FontWeight.bold, fontSize: 20),
-                    onChanged: (int newValue) {
+                    style: TextStyle(
+                        color: Colors.black,
+                        fontWeight: FontWeight.bold,
+                        fontSize: 20),
+                    onChanged: (int? newValue) {
                       setState(() {
-                        dropdownValueSize = newValue;
+                        dropdownValueSize = newValue!;
                       });
                     },
-                    items: <int>[20, 30, 40, 50, 70, 80, 90, 100].map<DropdownMenuItem<int>>((int value) {
+                    items: <int>[20, 30, 40, 50, 70, 80, 90, 100]
+                        .map<DropdownMenuItem<int>>((int value) {
                       return DropdownMenuItem<int>(
                         value: value,
                         child: Text(value.toString()),
@@ -347,7 +380,7 @@ class _Imprimir extends State<Imprimir> {
                 children: <Widget>[
                   SizedBox(
                     width: 150,
-                    child: RaisedButton(
+                    child: ElevatedButton(
                       child: Text(
                         'IMPRIMIR TEXTO',
                         style: TextStyle(fontSize: 12),
@@ -367,7 +400,7 @@ class _Imprimir extends State<Imprimir> {
                   ),
                   SizedBox(
                     width: 150,
-                    child: RaisedButton(
+                    child: ElevatedButton(
                       child: Text(
                         'IMPRIMIR IMAGEM',
                         style: TextStyle(fontSize: 12),
@@ -391,13 +424,17 @@ class _Imprimir extends State<Imprimir> {
                         icon: Icon(Icons.arrow_downward),
                         iconSize: 24,
                         elevation: 16,
-                        style: TextStyle(color: Colors.black, fontWeight: FontWeight.bold, fontSize: 20),
-                        onChanged: (int newValue) {
+                        style: TextStyle(
+                            color: Colors.black,
+                            fontWeight: FontWeight.bold,
+                            fontSize: 20),
+                        onChanged: (int? newValue) {
                           setState(() {
-                            dropdownBarHeight = newValue;
+                            dropdownBarHeight = newValue!;
                           });
                         },
-                        items: <int>[10, 40, 80, 120, 160, 200, 280, 320].map<DropdownMenuItem<int>>((int value) {
+                        items: <int>[10, 40, 80, 120, 160, 200, 280, 320]
+                            .map<DropdownMenuItem<int>>((int value) {
                           return DropdownMenuItem<int>(
                             value: value,
                             child: Text(value.toString()),
@@ -414,13 +451,17 @@ class _Imprimir extends State<Imprimir> {
                         icon: Icon(Icons.arrow_downward),
                         iconSize: 24,
                         elevation: 16,
-                        style: TextStyle(color: Colors.black, fontWeight: FontWeight.bold, fontSize: 20),
-                        onChanged: (int newValue) {
+                        style: TextStyle(
+                            color: Colors.black,
+                            fontWeight: FontWeight.bold,
+                            fontSize: 20),
+                        onChanged: (int? newValue) {
                           setState(() {
-                            dropdownBarWidth = newValue;
+                            dropdownBarWidth = newValue!;
                           });
                         },
-                        items: <int>[10, 40, 80, 120, 160, 200, 280, 320].map<DropdownMenuItem<int>>((int value) {
+                        items: <int>[10, 40, 80, 120, 160, 200, 280, 320]
+                            .map<DropdownMenuItem<int>>((int value) {
                           return DropdownMenuItem<int>(
                             value: value,
                             child: Text(value.toString()),
@@ -437,10 +478,13 @@ class _Imprimir extends State<Imprimir> {
                         icon: Icon(Icons.arrow_downward),
                         iconSize: 24,
                         elevation: 16,
-                        style: TextStyle(color: Colors.black, fontWeight: FontWeight.bold, fontSize: 20),
-                        onChanged: (String newValue) {
+                        style: TextStyle(
+                            color: Colors.black,
+                            fontWeight: FontWeight.bold,
+                            fontSize: 20),
+                        onChanged: (String? newValue) {
                           setState(() {
-                            dropdownBarType = newValue;
+                            dropdownBarType = newValue!;
                           });
                         },
                         items: <String>[
@@ -462,7 +506,7 @@ class _Imprimir extends State<Imprimir> {
               ),
               SizedBox(
                 width: 300,
-                child: RaisedButton(
+                child: ElevatedButton(
                   child: Text(
                     'IMPRIMIR BARCODE',
                     style: TextStyle(fontSize: 15),
@@ -481,7 +525,7 @@ class _Imprimir extends State<Imprimir> {
               ),
               SizedBox(
                 width: 300,
-                child: RaisedButton(
+                child: ElevatedButton(
                   child: Text(
                     'IMPRIMIR TODAS FUNÇÕES',
                     style: TextStyle(fontSize: 15),
