@@ -2,7 +2,7 @@ import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gertec/config_tef/operacaoRetorno.dart';
 import 'package:flutter_gertec/serviceTet.dart';
-import 'package:flutter_masked_text/flutter_masked_text.dart';
+import 'package:flutter_masked_text2/flutter_masked_text2.dart';
 
 class PageTef extends StatefulWidget {
   @override
@@ -41,9 +41,9 @@ class _PageTefState extends State<PageTef> {
   String tefSelecionado = "ger7"; //armazena a Tef escolhida
 
   // Altera o valor da opcao de habilitar impressao (true, false)
-  void alterarValorImpressao(bool newValue) => setState(
+  void alterarValorImpressao(bool? newValue) => setState(
         () {
-          habilitarImpressao = newValue;
+          habilitarImpressao = newValue!;
         },
       );
 
@@ -129,7 +129,7 @@ class _PageTefState extends State<PageTef> {
             ),
           ),
           actions: <Widget>[
-            FlatButton(
+            ElevatedButton(
               child: Text("Ok"),
               onPressed: () {
                 Navigator.pop(context);
@@ -160,7 +160,7 @@ class _PageTefState extends State<PageTef> {
             ),
           ),
           actions: <Widget>[
-            FlatButton(
+            ElevatedButton(
               child: Text("Ok"),
               onPressed: () {
                 Navigator.pop(context);
@@ -198,7 +198,7 @@ class _PageTefState extends State<PageTef> {
                 ],
               )),
           actions: <Widget>[
-            FlatButton(
+            ElevatedButton(
               child: Text("Ok"),
               onPressed: () {
                 Navigator.pop(context);
@@ -228,7 +228,7 @@ class _PageTefState extends State<PageTef> {
             ),
           ),
           actions: <Widget>[
-            FlatButton(
+            ElevatedButton(
               child: Text("Ok"),
               onPressed: () {
                 Navigator.pop(context);
@@ -241,19 +241,19 @@ class _PageTefState extends State<PageTef> {
   }
 
   //Marca o valor do tipo de Tef escolhido
-  void radioButtonChangeTef(String value) {
+  void radioButtonChangeTef(String? value) {
     setState(
       () {
-        tefSelecionado = value;
+        tefSelecionado = value!;
       },
     );
   }
 
   //Marca o valor do tipo de pagamento escolhido
-  void radioButtonChangePagamento(String value) {
+  void radioButtonChangePagamento(String? value) {
     setState(
       () {
-        tipoPagamentoSelecionado = value;
+        tipoPagamentoSelecionado = value!;
       },
     );
   }
@@ -476,7 +476,7 @@ class _PageTefState extends State<PageTef> {
     );
   }
 
-  Widget radioCheck(String text, String controll, Function onChange) {
+  Widget radioCheck(String text, String controll, Function(String?) onChange) {
     return SizedBox(
       height: 30,
       child: Radio(
@@ -491,7 +491,7 @@ class _PageTefState extends State<PageTef> {
     return Center(
       child: SizedBox(
         width: MediaQuery.of(context).size.width - 30,
-        child: RaisedButton(
+        child: ElevatedButton(
           onPressed: callback,
           child: Text(
             text,
