@@ -1,11 +1,11 @@
 class RetornoSat {
-  List<String> _retornoPipe; // Armazena o retorno em forma de lista String
-  String _operacao; // Armazena a operacao solicitada
-  bool retornoDiferente;
-  String _retornoPipeCompleto;
-  RetornoSat(String operacao, String retornopipe) {
+  List<String>? _retornoPipe; // Armazena o retorno em forma de lista String
+  String? _operacao; // Armazena a operacao solicitada
+  bool? retornoDiferente;
+  String? _retornoPipeCompleto;
+  RetornoSat(String? operacao, String? retornopipe) {
     this._retornoPipeCompleto = retornopipe;
-    this._retornoPipe = retornopipe.split("|");
+    this._retornoPipe = retornopipe?.split("|");
     this._operacao = operacao;
     /** 
       O resultado das operações solicitadas é uma String divida por Pipes, exemplo: “numeroSessao|EEEEE|mensagem|cod|mensagemSEFAZ”
@@ -13,7 +13,9 @@ class RetornoSat {
       Para solucionar este obstaculo foi elabarado uma condição que verifica se a operação solicitada é uma das três(que possuem maior discrepância de posições)
       Caso sim, é retornado uma posição diferente do mesmo valor, caso contrario é retornado a posição padrão
     **/
-    retornoDiferente = _operacao == 'AssociarSAT' || _operacao == 'EnviarTesteVendas' || _operacao == 'CancelarUltimaVenda';
+    retornoDiferente = _operacao == 'AssociarSAT' ||
+        _operacao == 'EnviarTesteVendas' ||
+        _operacao == 'CancelarUltimaVenda';
   }
   // Retorna a String divida por pipe sem formatações
   get getResultadoCompleto => _retornoPipeCompleto;
@@ -23,18 +25,19 @@ class RetornoSat {
 
   // Verifica se tem algum problema no retorno da operacao solicitada
   get getErroSat {
-    //* Se o tamanho da lista de informações obtidas do retorno [_retornoPipe] for igual <= 1, provavelmente ocorreu um erro.
-    if (_retornoPipe.length <= 1) {
+    //* Se o tamanho da lista de informações obtidas do retorno [_retornoPipe!] for igual <= 1, provavelmente ocorreu um erro.
+    if (_retornoPipe!.length <= 1) {
       // Verifica se houve um erro ao se conectar com o Sat
-      if (_retornoPipe[0] == 'Failed to find SAT device.') {
+      if (_retornoPipe![0] == 'Failed to find SAT device.') {
         return 'Dispositivo SAT não localizado';
       } else {
         // Caso não seja um erro de comunicação com o SAT retorna o erro que ocorreu
-        return _retornoPipe[0];
+        return _retornoPipe![0];
       }
     }
     //* Verifica se o código de ativação é invalido
-    if (getMensagem == "Codigo de ativação inválido" || getMensagem == "Codigo ativação inválido") {
+    if (getMensagem == "Codigo de ativação inválido" ||
+        getMensagem == "Codigo ativação inválido") {
       return getMensagem;
     }
     // Caso não exista nenhum erro
@@ -43,7 +46,7 @@ class RetornoSat {
 
   get getNumeroSessao {
     try {
-      return _retornoPipe[0];
+      return _retornoPipe![0];
     } catch (e) {
       return "";
     }
@@ -51,7 +54,7 @@ class RetornoSat {
 
   get getNumeroEEEE {
     try {
-      return _retornoPipe[1];
+      return _retornoPipe![1];
     } catch (e) {
       return "";
     }
@@ -59,8 +62,8 @@ class RetornoSat {
 
   get getMensagem {
     try {
-      if (retornoDiferente) return _retornoPipe[3];
-      return _retornoPipe[2];
+      if (retornoDiferente!) return _retornoPipe![3];
+      return _retornoPipe![2];
     } catch (e) {
       return "";
     }
@@ -69,7 +72,7 @@ class RetornoSat {
   // Dado somente presente na três funções(AssociarSAT, Enviar Teste Venda e Cancelar Ultima Venda)
   get getNumeroCCCC {
     try {
-      if (retornoDiferente) return _retornoPipe[2];
+      if (retornoDiferente!) return _retornoPipe![2];
       return "";
     } catch (e) {
       return "";
@@ -78,8 +81,8 @@ class RetornoSat {
 
   get getNumeroCod {
     try {
-      if (retornoDiferente) return _retornoPipe[4];
-      return _retornoPipe[3];
+      if (retornoDiferente!) return _retornoPipe![4];
+      return _retornoPipe![3];
     } catch (e) {
       return "";
     }
@@ -87,8 +90,8 @@ class RetornoSat {
 
   get getMensagemSefaz {
     try {
-      if (retornoDiferente) return _retornoPipe[5];
-      return _retornoPipe[4];
+      if (retornoDiferente!) return _retornoPipe![5];
+      return _retornoPipe![4];
     } catch (e) {
       return "";
     }
@@ -97,7 +100,7 @@ class RetornoSat {
   // Dado exclusivo do retorno da operação Ativar Sat
   get getCodigoCSR {
     try {
-      if (_operacao == 'AtivarSAT') return _retornoPipe[5];
+      if (_operacao == 'AtivarSAT') return _retornoPipe![5];
       return "";
     } catch (e) {
       return "";
@@ -107,7 +110,7 @@ class RetornoSat {
   // Dado exclusivo do retorno da operação Extrair Log
   get getLogBase64 {
     try {
-      if (_operacao == 'ExtrairLog') return _retornoPipe[5];
+      if (_operacao == 'ExtrairLog') return _retornoPipe![5];
       return "";
     } catch (e) {
       return "";
@@ -117,10 +120,11 @@ class RetornoSat {
   // Dado somente presente na três funções(EnviarTesteFim, Enviar Teste Venda e Cancelar Ultima Venda)
   get getArquivoCFeBase64 {
     try {
-      if (_operacao == 'EnviarTesteVendas' || _operacao == 'CancelarUltimaVenda') {
-        return _retornoPipe[6];
+      if (_operacao == 'EnviarTesteVendas' ||
+          _operacao == 'CancelarUltimaVenda') {
+        return _retornoPipe![6];
       } else if (_operacao == 'EnviarTesteFim') {
-        return _retornoPipe[5];
+        return _retornoPipe![5];
       } else {
         return "";
       }
@@ -132,10 +136,11 @@ class RetornoSat {
   // Dado somente presente na três funções(EnviarTesteFim, Enviar Teste Venda e Cancelar Ultima Venda)
   get getTimeStamp {
     try {
-      if (_operacao == 'EnviarTesteVendas' || _operacao == 'CancelarUltimaVenda') {
-        return _retornoPipe[7];
+      if (_operacao == 'EnviarTesteVendas' ||
+          _operacao == 'CancelarUltimaVenda') {
+        return _retornoPipe![7];
       } else if (_operacao == 'EnviarTesteFim') {
-        return _retornoPipe[6];
+        return _retornoPipe![6];
       } else {
         return "";
       }
@@ -147,7 +152,7 @@ class RetornoSat {
   // Dado exclusivo do retorno da operação TEnviarTesteFim
   get getNumDocFiscal {
     try {
-      if (_operacao == "EnviarTesteFim") return _retornoPipe[7];
+      if (_operacao == "EnviarTesteFim") return _retornoPipe![7];
       return "";
     } catch (e) {
       return "";
@@ -157,7 +162,9 @@ class RetornoSat {
   // Dado somente presente na três funções(EnviarTesteFim, Enviar Teste Venda e Cancelar Ultima Venda)
   get getChaveConsulta {
     try {
-      if (_operacao == 'EnviarTesteVendas' || _operacao == 'CancelarUltimaVenda' || _operacao == 'EnviarTesteFim') return _retornoPipe[8];
+      if (_operacao == 'EnviarTesteVendas' ||
+          _operacao == 'CancelarUltimaVenda' ||
+          _operacao == 'EnviarTesteFim') return _retornoPipe![8];
       return "";
     } catch (e) {
       return "";
@@ -167,8 +174,9 @@ class RetornoSat {
   // Dado somente presente nas funções(Enviar Teste Venda e Cancelar Ultima Venda)
   get getValorTotalCFe {
     try {
-      if (_operacao == 'EnviarTesteVendas' || _operacao == 'CancelarUltimaVenda')
-        return _retornoPipe[9];
+      if (_operacao == 'EnviarTesteVendas' ||
+          _operacao == 'CancelarUltimaVenda')
+        return _retornoPipe![9];
       else
         return "";
     } catch (e) {
@@ -179,8 +187,9 @@ class RetornoSat {
   // Dado somente presente nas funções(Enviar Teste Venda e Cancelar Ultima Venda)
   get getCPFCNPJValue {
     try {
-      if (_operacao == 'EnviarTesteVendas' || _operacao == 'CancelarUltimaVenda')
-        return _retornoPipe[10];
+      if (_operacao == 'EnviarTesteVendas' ||
+          _operacao == 'CancelarUltimaVenda')
+        return _retornoPipe![10];
       else
         return "";
     } catch (e) {
@@ -191,8 +200,9 @@ class RetornoSat {
   // Dado somente presente nas funções(Enviar Teste Venda e Cancelar Ultima Venda)
   get getAssinaturaQRCODE {
     try {
-      if (_operacao == 'EnviarTesteVendas' || _operacao == 'CancelarUltimaVenda')
-        return _retornoPipe[11];
+      if (_operacao == 'EnviarTesteVendas' ||
+          _operacao == 'CancelarUltimaVenda')
+        return _retornoPipe![11];
       else
         return "";
     } catch (e) {
@@ -202,15 +212,15 @@ class RetornoSat {
 
   get getEstadoDeOperacao {
     try {
-      if (_retornoPipe[27] == "0")
+      if (_retornoPipe![27] == "0")
         return "DESBLOQUEADO";
-      else if (_retornoPipe[27] == "1")
+      else if (_retornoPipe![27] == "1")
         return "BLOQUEADO SEFAZ";
-      else if (_retornoPipe[27] == "2")
+      else if (_retornoPipe![27] == "2")
         return "BLOQUEIO CONTRIBUINTE";
-      else if (_retornoPipe[27] == "3")
+      else if (_retornoPipe![27] == "3")
         return "BLOQUEIO AUTÔNOMO";
-      else if (_retornoPipe[27] == "4") return "BLOQUEIO PARA DESATIVAÇÃO";
+      else if (_retornoPipe![27] == "4") return "BLOQUEIO PARA DESATIVAÇÃO";
     } catch (e) {
       return "";
     }
@@ -218,67 +228,67 @@ class RetornoSat {
 
   get getNumeroSerieSat {
     try {
-      return _retornoPipe[5] + "\n";
+      return _retornoPipe![5] + "\n";
     } catch (e) {
       return "";
     }
   }
 
-  get getTipoLan => _retornoPipe[6] + "\n";
-  get getIpSat => _retornoPipe[7] + "\n";
-  get getMacSat => _retornoPipe[8] + "\n";
-  get getMascara => _retornoPipe[9] + "\n";
-  get getGateway => _retornoPipe[10] + "\n";
-  get getDns1 => _retornoPipe[11] + "\n";
-  get getDns2 => _retornoPipe[12] + "\n";
-  get getStatusRede => _retornoPipe[13] + "\n";
-  get getNivelBateria => _retornoPipe[14] + "\n";
-  get getMemoriaDeTrabalhoTotal => _retornoPipe[15] + "\n";
-  get getMemoriaDeTrabalhoUsada => _retornoPipe[16] + "\n";
+  get getTipoLan => _retornoPipe![6] + "\n";
+  get getIpSat => _retornoPipe![7] + "\n";
+  get getMacSat => _retornoPipe![8] + "\n";
+  get getMascara => _retornoPipe![9] + "\n";
+  get getGateway => _retornoPipe![10] + "\n";
+  get getDns1 => _retornoPipe![11] + "\n";
+  get getDns2 => _retornoPipe![12] + "\n";
+  get getStatusRede => _retornoPipe![13] + "\n";
+  get getNivelBateria => _retornoPipe![14] + "\n";
+  get getMemoriaDeTrabalhoTotal => _retornoPipe![15] + "\n";
+  get getMemoriaDeTrabalhoUsada => _retornoPipe![16] + "\n";
   get getDataHora =>
-      _retornoPipe[17].substring(6, 8) +
+      _retornoPipe![17].substring(6, 8) +
       "/" +
-      _retornoPipe[17].substring(4, 6) +
+      _retornoPipe![17].substring(4, 6) +
       "/" +
-      _retornoPipe[17].substring(0, 4) +
+      _retornoPipe![17].substring(0, 4) +
       "  " +
-      _retornoPipe[17].substring(8, 10) +
+      _retornoPipe![17].substring(8, 10) +
       ":" +
-      _retornoPipe[17].substring(10, 12) +
+      _retornoPipe![17].substring(10, 12) +
       ":" +
-      _retornoPipe[17].substring(12, 14) +
+      _retornoPipe![17].substring(12, 14) +
       "\n";
 
-  get getVersao => _retornoPipe[18] + "\n";
-  get getVersaoLeiaute => _retornoPipe[19] + "\n";
-  get getUltimoCfeEmitido => _retornoPipe[20] + "\n";
-  get getPrimeiroCfeMemoria => _retornoPipe[21] + "\n";
-  get getUltimoCfeMemoria => _retornoPipe[22] + "\n";
+  get getVersao => _retornoPipe![18] + "\n";
+  get getVersaoLeiaute => _retornoPipe![19] + "\n";
+  get getUltimoCfeEmitido => _retornoPipe![20] + "\n";
+  get getPrimeiroCfeMemoria => _retornoPipe![21] + "\n";
+  get getUltimoCfeMemoria => _retornoPipe![22] + "\n";
   get getUltimaTransmissaoSefazDataHora =>
-      _retornoPipe[23].substring(6, 8) +
+      _retornoPipe![23].substring(6, 8) +
       "/" +
-      _retornoPipe[23].substring(4, 6) +
+      _retornoPipe![23].substring(4, 6) +
       "/" +
-      _retornoPipe[23].substring(0, 4) +
+      _retornoPipe![23].substring(0, 4) +
       " " +
-      _retornoPipe[23].substring(8, 10) +
+      _retornoPipe![23].substring(8, 10) +
       ":" +
-      _retornoPipe[23].substring(10, 12) +
+      _retornoPipe![23].substring(10, 12) +
       ":" +
-      _retornoPipe[23].substring(12, 14) +
+      _retornoPipe![23].substring(12, 14) +
       "\n";
 
   get getUltimaComunicacaoSefazData =>
-      _retornoPipe[24].substring(6, 8) +
+      _retornoPipe![24].substring(6, 8) +
       "/" +
-      _retornoPipe[24].substring(4, 6) +
+      _retornoPipe![24].substring(4, 6) +
       "/" +
-      _retornoPipe[24].substring(0, 4) +
+      _retornoPipe![24].substring(0, 4) +
       " " +
-      _retornoPipe[24].substring(8, 10) +
+      _retornoPipe![24].substring(8, 10) +
       ":" +
-      _retornoPipe[24].substring(10, 12) +
+      _retornoPipe![24].substring(10, 12) +
       ":" +
-      _retornoPipe[24].substring(12, 14) +
+      _retornoPipe![24].substring(12, 14) +
       "\n";
 }
