@@ -20,7 +20,12 @@ Map<String, dynamic> _formatarInfoRecebida(myjson) {
 }
 
 class TefService {
-  TefService({String valor, String tipoPagamento, int quantParcelas, bool habilitarImpressao, String ip}) {
+  TefService(
+      {String? valor,
+      String? tipoPagamento,
+      int? quantParcelas,
+      bool? habilitarImpressao,
+      String? ip}) {
     this._ipConfig = ip;
     this._valorVenda = valor;
     this._tipoPagamento = tipoPagamento;
@@ -29,12 +34,12 @@ class TefService {
   }
 
   final _platform = const MethodChannel('samples.flutter.dev/gedi');
-  String _valorVenda;
-  String _tipoPagamento;
-  int _quantParcelas;
-  bool _habilitarImpressao;
-  String _ipConfig;
-  String _tipoParcelamento;
+  String? _valorVenda;
+  String? _tipoPagamento;
+  int? _quantParcelas;
+  bool? _habilitarImpressao;
+  String? _ipConfig;
+  String? _tipoParcelamento;
 
   final String _ger7ApiVersion = "1.04";
   final String _ger7SemParcelamento = "0";
@@ -44,8 +49,8 @@ class TefService {
   final String _ger7HabilitaImpressao = "1";
 
 //Metodos Get
-  String get getIpConfig => _ipConfig;
-  String get getValorVenda => _valorVenda;
+  String get getIpConfig => _ipConfig!;
+  String get getValorVenda => _valorVenda!;
 
   //Retorna uma lista onde o Index 0 == TipoPagamento Ger 7 e o 1 == TipoPagamento M-Sitef (Caso exista esse tipo de pagamento no Tef)
   List<String> get getTipoPagamento {
@@ -60,8 +65,8 @@ class TefService {
     }
   }
 
-  int get getQuantParcelas => _quantParcelas;
-  bool get getImpressaoHabilitada => _habilitarImpressao;
+  int get getQuantParcelas => _quantParcelas!;
+  bool get getImpressaoHabilitada => _habilitarImpressao!;
   get getTipoParcelamento => this._tipoParcelamento;
 
 //Metodos Set
@@ -159,8 +164,8 @@ class TefService {
   /// Realiza a formatação para realização de venda MsiTef.
   ///
   /// Retorna um json, que vai ser enviado para a Tef MsiTef.
-  Map<String, String> get _formatarPametrosVendaMsiTef {
-    Map<String, String> mapMsiTef = Map();
+  Map<String, String?>? get _formatarPametrosVendaMsiTef {
+    Map<String, String?>? mapMsiTef = Map();
     mapMsiTef["empresaSitef"] = "00000000";
     mapMsiTef["enderecoSitef"] = this.getIpConfig;
     mapMsiTef["operador"] = "0001";
@@ -206,8 +211,8 @@ class TefService {
   ///Realiza a formatação para realização de cancelamento MsiTef.
   ///
   /// Retorna um map <String, String>, que vai ser enviado para a MsiTef.
-  Map<String, String> get _formatarPametrosCancelamentoMsiTef {
-    Map<String, String> mapMsiTef = Map();
+  Map<String, String?>? get _formatarPametrosCancelamentoMsiTef {
+    Map<String, String?>? mapMsiTef = Map();
     mapMsiTef["empresaSitef"] = "00000000";
     mapMsiTef["enderecoSitef"] = this.getIpConfig;
     mapMsiTef["operador"] = "0001";
@@ -235,8 +240,8 @@ class TefService {
   /// Realiza a formatação para realização de configuração de funções MsiTef.
   ///
   /// Retorna um map <String, String>, que vai ser enviado para a Tef Ger 7.
-  Map<String, String> get _formatarPametrosFuncoesMsiTef {
-    Map<String, String> mapMsiTef = Map();
+  Map<String, String?>? get _formatarPametrosFuncoesMsiTef {
+    Map<String, String?>? mapMsiTef = Map();
     mapMsiTef["empresaSitef"] = "00000000";
     mapMsiTef["enderecoSitef"] = this.getIpConfig;
     mapMsiTef["operador"] = "0001";
@@ -265,8 +270,8 @@ class TefService {
   /// Realiza a formatação para realização de reimpressão MsiTef.
   ///
   /// Retorna um map <String, String>, que vai ser enviado para a Tef Ger
-  Map<String, String> get _formatarPametrosReimpressaoMsiTef {
-    Map<String, String> mapMsiTef = Map();
+  Map<String, String?>? get _formatarPametrosReimpressaoMsiTef {
+    Map<String, String?>? mapMsiTef = Map();
     mapMsiTef["empresaSitef"] = "00000000";
     mapMsiTef["enderecoSitef"] = this.getIpConfig;
     mapMsiTef["operador"] = "0001";
@@ -292,7 +297,8 @@ class TefService {
   // Realiza as funções da Tef, tem como retorno um objeto dynamic que pode ser atributi a [RetornoMsiTef] ou [RetornoGer7]
   // Recebe como parâmetros uma String que está relacionada a ação que deseja ser invocada e uma String relacionado a tef utilizada (ger7,msitef)
   // As ações possiveis são: venda, cancelamento, reimpressao, funcoes (Os valores devem ser escritos exatamente como o demonstrado)
-  Future<dynamic> enviarParametrosTef({@required String tipoAcao, @required String tipoTef}) async {
+  Future<dynamic> enviarParametrosTef(
+      {required String tipoAcao, required String tipoTef}) async {
     var retornoTef;
     var myjson;
     var parametroFormatado;
